web/comori-od-adminsvc/requirements.txt
cffi
falcon
gunicorn
gevent
python-rapidjson
pyyaml
requests
simplejson
unidecode
bs4
readtime
prettytable